Moved and edited isRequestingUserAnAdminOnCorpus function to test to see if requesting user is admin before allowing user to change roles.
