+++ lib/corpus.js
@@ -16,7 +16,7 @@ var admin = "admin";
 var createNewCorpus = function(req, callback) {
   /* TODO not sure if where the serverCode is coming from, it theoretically should be in the body... */
   var serverLabel = req.body.serverCode || req.serverCode;
-  if(!serverLabel){
+  if (!serverLabel) {
     serverLabel = req.body.serverLabel;
   }
   var username = req.body.username;
@@ -145,7 +145,7 @@ var updateRoles = function(req, callback) {
 
   /* TODO not sure where the serverCode is coming from, it theoretically should be in the body not in the req... */
   var serverLabel = req.body.serverCode || req.serverCode;
-  if(!serverLabel){
+  if (!serverLabel) {
     serverLabel = req.body.serverLabel;
   }
   var couchConnection = this.getCouchConnectionFromServerCode(serverLabel);
@@ -154,79 +154,89 @@ var updateRoles = function(req, callback) {
   var username = req.body.userRoleInfo.usernameToModify;
 
   // Check to make sure requesting user is admin; if not, abort
+  var requestingUser = req.body.username;
+  isRequestingUserAnAdminOnCorpus(req, requestingUser, couchConnection, function(error, result, messages) {
+    if (error) {
+      console.log("ERROR HERE!");
+      console.log(error);
+      return callback(true, username, {
+        "message": "You do not have permission to update user roles for this corpus. You must be an admin."
+      });
+    } else {
+      console.log(new Date() + " User " + requestingUser + " is admin and can modify permissions on " + couchConnection.pouchname);
 
+      if (req.body.userRoleInfo.removeUser && req.body.userRoleInfo.removeUser == true) {
+        console.log("Removing permissions for " + username + " from " + couchConnection.pouchname);
+        removeUserFromDb(couchConnection, username);
+        return callback(false, username, {});
+      }
 
-  //Seriously?! just take them out of the corpus team with out making sure the requester has rights?
-  if (req.body.userRoleInfo.removeUser && req.body.userRoleInfo.removeUser == true) {
-    console.log("Removing permissions for " + username + " from " + couchConnection.pouchname);
-    removeUserFromDb(couchConnection, username);
-    return callback(false, username, {});
-  }
+      var roles = [];
+      if (req.body.userRoleInfo.admin == true) {
+        roles.push(couchConnection.pouchname + "_" + admin);
+      }
+      if (req.body.userRoleInfo.reader == true) {
+        roles.push(couchConnection.pouchname + "_" + collaborator);
+      }
+      if (req.body.userRoleInfo.writer == true) {
+        roles.push(couchConnection.pouchname + "_" + contributor);
+      }
 
-  var roles = [];
-  if (req.body.userRoleInfo.admin == true) {
-    roles.push(couchConnection.pouchname + "_" + admin);
-  }
-  if (req.body.userRoleInfo.reader == true) {
-    roles.push(couchConnection.pouchname + "_" + collaborator);
-  }
-  if (req.body.userRoleInfo.writer == true) {
-    roles.push(couchConnection.pouchname + "_" + contributor);
-  }
+      // Check to see if user exists before adding roles
+      var connect = couchConnection.protocol + couch_keys.username + ":" + couch_keys.password + "@" + couchConnection.domain + ":" + couchConnection.port + "/_users";
 
-  // Check to see if user exists before adding roles
-  var connect = couchConnection.protocol + couch_keys.username + ":" + couch_keys.password + "@" + couchConnection.domain + ":" + couchConnection.port + "/_users";
+      var db = require('nano')(connect);
+      var userid = 'org.couchdb.user:' + username;
 
-  var db = require('nano')(connect);
-  var userid = 'org.couchdb.user:' + username;
-
-  db.get(userid, function(err, body) {
-    if (!err) {
-      // User exists on server, so adding new user roles
-      addRoleToUserInfo(couchConnection, username, roles);
-
-      // Add user doc in corpus if it does not already exist
-      var corpusConnect = couchConnection.protocol + couch_keys.username + ":" + couch_keys.password + "@" + couchConnection.domain + ":" + couchConnection.port + "/" + couchConnection.pouchname;
-
-      var corpusDb = require('nano')(corpusConnect);
-      corpusDb.get(username, function(err, body) {
+      db.get(userid, function(err, body) {
         if (!err) {
-          // Do nothing if user already exists
-        } else {
-          // Add new user to DB
-          // Get gravatar hash
-          var zUserInfoDbConnect = couchConnection.protocol + couch_keys.username + ":" + couch_keys.password + "@" + couchConnection.domain + ":" + couchConnection.port + "/zfielddbuserscouch";
+          // User exists on server, so adding new user roles
+          addRoleToUserInfo(couchConnection, username, roles);
 
-          var zUserInfoDb = require('nano')(zUserInfoDbConnect);
+          // Add user doc in corpus if it does not already exist
+          var corpusConnect = couchConnection.protocol + couch_keys.username + ":" + couch_keys.password + "@" + couchConnection.domain + ":" + couchConnection.port + "/" + couchConnection.pouchname;
 
-          zUserInfoDb.get(username, function(err, body) {
+          var corpusDb = require('nano')(corpusConnect);
+          corpusDb.get(username, function(err, body) {
             if (!err) {
-              var userDoc = {
-                "gravatar": body.gravatar,
-                "username": username,
-                "authUrl": couchConnection.authUrl,
-                "id": username,
-                "collection": "users"
-              };
+              // Do nothing if user already exists
+            } else {
+              // Add new user to DB
+              // Get gravatar hash
+              var zUserInfoDbConnect = couchConnection.protocol + couch_keys.username + ":" + couch_keys.password + "@" + couchConnection.domain + ":" + couchConnection.port + "/zfielddbuserscouch";
 
-              // Save new user record to corpus
-              corpusDb.insert(userDoc, username, function(err, body) {
+              var zUserInfoDb = require('nano')(zUserInfoDbConnect);
+
+              zUserInfoDb.get(username, function(err, body) {
                 if (!err) {
-                  console.log("Successfully added new user doc to corpus.");
+                  var userDoc = {
+                    "gravatar": body.gravatar,
+                    "username": username,
+                    "authUrl": couchConnection.authUrl,
+                    "id": username,
+                    "collection": "users"
+                  };
+
+                  // Save new user record to corpus
+                  corpusDb.insert(userDoc, username, function(err, body) {
+                    if (!err) {
+                      console.log("Successfully added new user doc to corpus.");
+                    } else {
+                      console.log(err);
+                    }
+                  });
                 } else {
                   console.log(err);
                 }
               });
-            } else {
-              console.log(err);
             }
           });
+
+          return callback(false, username, {});
+        } else {
+          return callback(true, username, {});
         }
       });
-
-      return callback(false, username, {});
-    } else {
-      return callback(true, username, {});
     }
   });
 };
@@ -306,6 +316,62 @@ var removeUserFromDb = function(couchConnection, username) {
     }
   });
 };
+
+/*
+ * Ensures the requesting user to make the permissions
+ * modificaitons, can be used for any corpus operations which require admin privildages.
+ */
+var isRequestingUserAnAdminOnCorpus = function(req, requestingUser, dbConn, done) {
+  if (!dbConn) {
+    return done({
+      error: "Client didn't define the database connection."
+    }, null, {
+      message: "There was a problem deciding if you have permission to do this."
+    });
+  }
+
+  /*
+   * Check to see if the user is an admin on the corpus
+   */
+  if (dbConn.domain.indexOf("iriscouch") > -1) {
+    dbConn.port = "6984";
+  }
+  var nanoforpermissions = require('nano')(dbConn.protocol + couch_keys.username + ':' + couch_keys.password + '@' + dbConn.domain + ':' + dbConn.port + dbConn.path);
+  var usersdb = nanoforpermissions.db.use("_users");
+  usersdb.get("org.couchdb.user:" + requestingUser, function(error, result) {
+    if (error) {
+      return done({
+        error: "User " + requestingUser + " couldn't be found on this server"
+      }, null, {
+        message: "There was a problem deciding if you have permission to do this."
+      });
+
+    } else {
+      var userIsAdminOnTeam = false;
+
+      if (result) {
+        var userroles = result.roles;
+        userroles.forEach(function(role) {
+          if (role.indexOf(dbConn.pouchname + "_" + admin) > -1) {
+            userIsAdminOnTeam = true;
+            return done(null, requestingUser, null);
+          }
+        });
+      }
+      /* Catch all */
+      if (userIsAdminOnTeam == false) {
+        return done({
+          error: "User was found but didnt have permission."
+        }, null, {
+          info: "User does not have permission to perform this action."
+        });
+      }
+    }
+  });
+};
+
+module.exports.isRequestingUserAnAdminOnCorpus = isRequestingUserAnAdminOnCorpus;
+
 
 var getCouchConnectionFromServerCode = function(serverCode) {
   serverCode = serverCode.toLowerCase();
@@ -388,7 +454,7 @@ var createPrivateCorpusDoc = function(username, corpusTitle, couchConnection, po
   newPrivateCorpus.confidential.secretkey = uuid.v4();
   newPrivateCorpus.timestamp = Date.now();
   newPrivateCorpus.sessionFields = JSON.parse(JSON.stringify(session_default_template)).sessionFields;
-  for(var field in newPrivateCorpus.sessionFields){
+  for (var field in newPrivateCorpus.sessionFields) {
     newPrivateCorpus.sessionFields[field].value = "";
     newPrivateCorpus.sessionFields[field].mask = "";
   }
