+++ lib/userauthentication.js
@@ -443,53 +443,6 @@ module.exports.authenticateUser = function(username, password, req, done) {
 
 /*
  * Ensures the requesting user to make the permissions
- * modificaitons, can be used for any corpus operations which require admin privildages. 
- */
-module.exports.isRequestingUserAnAdminOnCorpus = function(req, requestingUser, dbConn, done){
-  if (!dbConn) {
-    return done({
-      error: "Client didn't define the database connection."
-    }, null, {
-      message: "There was a problem deciding if you have permission to do this."
-    });
-  }
-
-  /*
-   * Check to see if the user is an admin on the corpus
-   */
-  if (dbConn.domain.indexOf("iriscouch") > -1) {
-    dbConn.port = "6984";
-  }
-  var nanoforpermissions = require('nano')(dbConn.protocol + couch_keys.username + ':' + couch_keys.password + '@' + dbConn.domain + ':' + dbConn.port + dbConn.path);
-  var usersdb = nanoforpermissions.db.use("_users");
-  usersdb.get("org.couchdb.user:" + requestingUser, function(error, result) {
-    if (error) {
-      return done({
-        error: "User " + requestingUser + " couldn't be found on this server"
-      }, null, {
-        message: "There was a problem deciding if you have permission to do this."
-      });
-
-    } else {
-      if (result) {
-        var userroles = result.roles;
-        var userIsAdminOnTeam = false;
-        userroles.forEach(function(role) {
-          if (role.indexOf(dbConn.pouchname + "_" + admin) > -1) {
-            userIsAdminOnTeam = true;
-            return done(null, requestingUser, null);
-          }
-        });
-      }
-    }
-  });
-
-  /* Catch all */
-  return done({error: "User was found but didnt have permission."}, null, {info: "User does not have permission to perform this action."});
-};
-
-/*
- * Ensures the requesting user to make the permissions
  * modificaitons. Then adds the role to the user if they exist
  */
 module.exports.addRoleToUser = function(req, done) {
@@ -504,10 +457,10 @@ module.exports.addRoleToUser = function(req, done) {
   }
 
 
-  isRequestingUserAnAdminOnCorpus(req, requestingUser, dbConn, function(error, result, messages){
-    if(error){
+  corpus.isRequestingUserAnAdminOnCorpus(req, requestingUser, dbConn, function(error, result, messages) {
+    if (error) {
       return done(error, result, messages);
-    } else{
+    } else {
       console.log(new Date() + " User " + requestingUser + " is admin and can modify permissions on " + dbConn.pouchname + " : " + util.inspect(role));
 
       var rolesSimpleStrings = req.body.roles;
@@ -534,8 +487,6 @@ module.exports.addRoleToUser = function(req, done) {
 
     }
   });
-
-
 };
 
 /*
