+++ service.js
@@ -247,7 +247,7 @@ app.get('/addroletouser', function(req, res, next) {
 });
 
 /**
- * Responds to requests for adding a corpus to a user, if successful replies with the pouchname of the new corpus in a string and a corpusaded = true 
+ * Responds to requests for adding a corpus to a user, if successful replies with the pouchname of the new corpus in a string and a corpusaded = true
  */
 app.post('/newcorpus', function(req, res, next) {
   authenticationfunctions.authenticateUser(req.body.username, req.body.password, req, function(err, user, info) {
@@ -311,7 +311,7 @@ app.post('/updateroles', function(req, res, next) {
       corpus.updateRoles(req, function(err, roles, info) {
         if (err) {
           console.log(new Date() + " There was an error in corpus.updateRoles\n");
-          returndata.userFriendlyErrors = ["There was an error updating the user roles.\nUser " + roles + " does not exist."];
+          returndata.userFriendlyErrors = [info.message];
         }
         if (!roles) {
           returndata.userFriendlyErrors = ["There was an error updating the user roles."];
